refactor(asmdb): insnFormat.tsx -> insnFormatName.tsx
